User authentication/session management
Now all users are able to register with their password hashed to the database. Users can login and request will compare their entered password to the hashed password in the database. Additionally i have added session management which creates a Cookie with user authentication so users dont have to authenticate every time. I have also create a method to check for i user is authenticated, but i have not used it on any routes for the ease of development now.

Top navigation bar will reflect the changes based on if user is logged in or not. For users that are not logged in it will show 'register' button for users that are not logged in it will show 'log out' button