SQL Files for testing.
Made 3 good SQL files for testing and demo purposes. For Demo and most of test cases, use the "FULL" DB. The empty one doesn't have ingredients or recipes, and the partial one doesn't have enough ingredients to cook anything, but has a bunch of partial matches for cook it. For all of these files, when testing, make sure you manually change the system data on your computer to 4/18/19.